# docs/swagger.yaml
basePath: /
definitions:
...
  service.CreateUserInput:
    properties:
      email:
        type: string
      facility_id:
        type: string
      full_name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
  utils.APIError:
    properties:
      errors:
        items:
          $ref: '#/definitions/utils.ErrorDetail'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  utils.APIResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Meta'
      success:
        type: boolean
    type: object
  utils.ErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  utils.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
info:
  contact: {}
  description: Uganda MoH HMIS MCH 007 Integrated Family Planning Register
  title: HMIS MCH 007 – Integrated FP Register API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      - description: Filter by facility
        in: query
        name: facility_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - BearerAuth: []
      summary: Create a user
      tags:
      - users
  /api/v1/users/{id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /api/v1/users/{id}/deactivate:
    patch:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Deactivate user
      tags:
      - users
  /api/v1/users/import:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 'CSV file with columns: full_name,email,password,role,facility_id'
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Bulk import users from CSV
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"